+++ kernel/posix/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(kernel PRIVATE posix/pthread_common.c)
+target_sources(kernel PRIVATE posix/pthread_cond.c)
+target_sources(kernel PRIVATE posix/pthread_mutex.c)
+target_sources(kernel PRIVATE posix/pthread_barrier.c)
